\name{eggprod-package}
\alias{eggprod-package}
\alias{eggprod}
\docType{package}
\title{
traditional egg production estimate
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab eggprod\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-01-20\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Colin Millar based on the fortran code by Rob Fryer
Maintainer: Colin Millar <millar@marlab.ac.uk>
}
\references{
reference to document of robs
}
\keyword{ package }
\seealso{
}
\examples{
#
get_dates("2007")
}
